Add simple Twitter plugin for sending tweets (#103)
Super basic twitter plugin for Agreement Engine to use. It can only send tweets from the bot account that the app was created for. The plugin also emits events for tweets on the bots timeline. In the future, we can extend it with oauth support, which would allow communities to select their own Twitter account that they want it to act on behalf of. See docs for setup instructions. The action exposed is twitter.send-tweet.